version 2.0.4

+++ src/Api/Evolution/Device.php
@@ -590,6 +590,10 @@ class Device
 
          try {
 
+            if($httpCode == 404){
+               return true;
+            }
+
             if ($httpCode != 200) {
                ExceptionError::setError($httpCode, json_encode(['type' => 'Api response', 'class' => 'Api\Evolution\Device', 'method' => 'delete', 'message' => $response]));
                return NULL;
@@ -601,6 +605,12 @@ class Device
             }
 
             $json = json_decode($response);
+
+            if(isset($json->status)){
+               if($json->status == 404){
+                  return true;
+               }
+            }
 
             if (isset($json->status)) {
                if ($json->status == 'SUCCESS') {
